Replaced app.js with Golang based main.exe
Replaced Nodejs based app.js with Golang based executable main.exe. This is so that technicians no longer have to build an entire Nodejs environment just to run the code. It is all now bundled nicely in the executable which is easily run by anyone. Deleted all Nodejs related files.